Rewrote most of the parser and added full suport for repeat bars. The parser can now parse remembering rain. Also changed most of the draws to accept floats. Added size left and right padding to the ITabNotation interface and implemented them across the model classes.

+++ src/main/java/ca/yorku/cse2311/tab2pdf/model/IDrawable.java
@@ -12,6 +12,6 @@ import com.itextpdf.text.pdf.PdfWriter;
  */
 public interface IDrawable {
 
-    public void draw(int staveNumber, int lineNumber, int xCoordinate, PdfWriter writer);
+    public void draw(int staveNumber, int lineNumber, float xCoordinate, PdfWriter writer);
 
 }
